requirements.txt
mysql_connector_repackaged==0.3.1
natsort==8.4.0
nibabel==4.0.1
numpy==1.17.4
pandas==1.4.3
setuptools==45.2.0
